Update CakeAliasCategory to fit more with built-in aliases
