+++ src/Cake.Incubator/Cake.Incubator.csproj
@@ -60,6 +60,7 @@
     <Compile Include="DotNetCoreTestExtensions.cs" />
     <Compile Include="EnvironmentExtensions.cs" />
     <Compile Include="FileExtensions.cs" />
+    <Compile Include="GlobbingExtensions.cs" />
     <Compile Include="LoggingExtensions.cs" />
     <Compile Include="ProjectTypes.cs" />
     <Compile Include="StringExtensions.cs" />
