+++ src/Cake.Incubator/FileExtensions.cs
@@ -34,68 +34,6 @@ namespace Cake.Incubator
         }
 
         /// <summary>
-        /// Returns files in the same directory that have the same file name but different extensions
-        /// </summary>
-        /// <param name="context">the cake context</param>
-        /// <param name="files">the files to return matches for</param>
-        /// <returns>a list of matching files</returns>
-        /// <example>
-        /// Locates files with the same name in the same directory, but different extensions.
-        /// The .pdb to your .dll as it were.
-        /// <code>
-        /// // /output/file.dll
-        /// // /output/file.xml
-        /// // /output/file.pdb
-        /// // /output/another.dll
-        ///
-        /// IEnumerable&lt;FilePath&gt; matchingFiles = GetMatchingFiles(new FilePath("/output/file.dll"));
-        ///
-        /// matchingFiles[0]; // /output/file.xml
-        /// matchingFiles[1]; // /output/file.pdb
-        /// </code>
-        /// </example>
-        [CakeMethodAlias]
-        public static IEnumerable<IFile> GetMatchingFiles(this ICakeContext context, IEnumerable<FilePath> files)
-        {
-            return files.SelectMany(
-                x =>
-                    context.FileSystem.GetDirectory(x.GetDirectory())
-                        .GetFiles($"{x.GetFilenameWithoutExtension()}.*", SearchScope.Current)
-                        .Where(f => !f.Path.Equals(x)));
-        }
-
-        /// <summary>
-        /// Gets FilePaths using glob patterns
-        /// </summary>
-        /// <param name="context">the cake context</param>
-        /// <param name="patterns">the glob patterns</param>
-        /// <returns>the files matching the glob patterns</returns>
-        /// <example>
-        /// Locates files with the same name in the same directory, but different extensions.
-        /// The .pdb to your .dll as it were.
-        /// <code>
-        /// // /output/file.dll
-        /// // /output/file.xml
-        /// // /output/file.pdb
-        /// // /output/another.dll
-        ///
-        /// IEnumerable&lt;FilePath&gt; matchingFiles = GetMatchingFiles(new FilePath("/output/file.dll"));
-        ///
-        /// matchingFiles[0]; // /output/file.xml
-        /// matchingFiles[1]; // /output/file.pdb
-        /// </code>
-        /// </example>
-        [CakeMethodAlias]
-        public static FilePathCollection GetFiles(this ICakeContext context, params string[] patterns)
-        {
-            return
-                patterns.Aggregate(
-                    new FilePathCollection(new PathComparer(context.Environment)),
-                    (current, pattern) =>
-                        current + context.Globber.GetFiles(pattern));
-        }
-
-        /// <summary>
         /// Gets the output assembly paths for solution or project files, for a specific build configuration
         /// </summary>
         /// <param name="context">the cake context</param>
@@ -201,6 +139,7 @@ namespace Cake.Incubator
         /// <param name="source">the source file</param>
         /// <param name="destination">the destination file path</param>
         [CakeMethodAlias]
+        [CakeAliasCategory("Move")]
         [Obsolete("Use Cake.Common.IO.MoveFile instead")]
         public static void Move(this ICakeContext context, FilePath source, FilePath destination)
         {
@@ -220,6 +159,7 @@ namespace Cake.Incubator
         /// <param name="source"></param>
         /// <param name="destination"></param>
         [CakeMethodAlias]
+        [CakeAliasCategory("Move")]
         [Obsolete("Use Cake.Common.IO.MoveFiles instead")]
         public static void Move(this ICakeContext context, IEnumerable<FilePath> source, DirectoryPath destination)
         {
